+++ src/light_compressor/compressor_method.py
@@ -9,6 +9,7 @@ class CompressionMethod(Enum):
     LZ4 = 0x82
     ZSTD = 0x90
     GZIP = 0x99
+    SNAPPY = 0x9f
 
     @property
     def method(self) -> str:
@@ -22,12 +23,14 @@ def auto_detector(fileobj: BufferedReader) -> CompressionMethod:
     Warning!!! Not work with stream objects!!!"""
 
     pos = fileobj.tell()
-    signature = fileobj.read(4)
+    signature = fileobj.read(10)
     fileobj.seek(pos)
 
-    if signature == b"\x04\"M\x18":
+    if signature == b"\xff\x06\x00\x00sNaPpY":
+        return CompressionMethod.SNAPPY
+    if signature[:4] == b"\x04\"M\x18":
         return CompressionMethod.LZ4
-    if signature == b"(\xb5/\xfd":
+    if signature[:4] == b"(\xb5/\xfd":
         return CompressionMethod.ZSTD
     if signature[:2] == b"\x1f\x8b":
         return CompressionMethod.GZIP