+++ src/light_compressor/compressors/__init__.py
@@ -2,11 +2,13 @@
 
 from .gzip import GZIPCompressor
 from .lz4 import LZ4Compressor
+from .snappy import SNAPPYCompressor
 from .zstd import ZSTDCompressor
 
 
 __all__ = (
     "GZIPCompressor",
     "LZ4Compressor",
+    "SNAPPYCompressor",
     "ZSTDCompressor",
 )