+++ src/light_compressor/decompressors/__init__.py
@@ -3,6 +3,7 @@
 from .decompress_reader import DecompressReader
 from .gzip import GZIPDecompressor
 from .lz4 import LZ4Decompressor
+from .snappy import SNAPPYDecompressor
 from .zstd import ZSTDDecompressor
 
 
@@ -10,5 +11,6 @@ __all__ = (
     "DecompressReader",
     "GZIPDecompressor",
     "LZ4Decompressor",
+    "SNAPPYDecompressor",
     "ZSTDDecompressor",
 )